paddleocr-api/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
paddlepaddle>=2.6.0
paddleocr>=2.7.0
Pillow>=10.0.0
numpy>=1.24.0
